Things done
Removed SIMD vector and functions, because it is either slower or the same speed as Vec3.
Fixed castRay function
Added the ability to place and remove blocks.
Added simple shadows
Moved code from main to camera
Added P functions to world (functions that check if pos is inside the world or chunk)

TODO: clean up main